ai-engine/scraper.py
import requests
from bs4 import BeautifulSoup
import asyncio

# --- LEARNING NOTE: Scraping Basics ---
# 1. requests: Used to download the HTML content of a webpage.
# 2. BeautifulSoup: Used to "parse" the HTML and find specific tags (like <h1>, <div>, or <a>).
# 3. Selenium/Playwright (Advanced): Used for sites that use React/Vue and need a real browser.
# ---

async def scrape_jobs(keyword, country):
    """
    Simulates scraping a job board.
    In a real scenario, we would target LinkedIn, Indeed, etc.
    """
    print(f"--- Starting scrape for {keyword} in {country} ---")
    
    # Example Target (Mocking a search URL)
    url = f"https://www.google.com/search?q={keyword}+jobs+in+{country}+visa+sponsorship"
    headers = {
        "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
    }

    try:
        # We use loop.run_in_executor to run the blocking 'requests' call in a thread
        # This keeps our 'async' function from blocking the whole program.
        loop = asyncio.get_event_loop()
        response = await loop.run_in_executor(None, lambda: requests.get(url, headers=headers))
        
        if response.status_code == 200:
            soup = BeautifulSoup(response.text, 'html.parser')
            
            # This is where the magic happens:
            # We look for specific patterns in the text.
            # Real scraping requires fine-tuning for each website's HTML.
            print("Successfully fetched page content.")
            
            # Returning dummy data for now to keep it deterministic
            return [
                {
                    "title": f"Staff {keyword}",
                    "company": "Berlin Tech Group",
                    "location": "Berlin, Germany",
                    "sponsorship": "Confirmed",
                    "summary": "Full sponsorship provided for senior candidates.",
                    "url": "https://berlin-tech.com/jobs/1"
                }
            ]
        else:
            print(f"Failed to fetch. Status code: {response.status_code}")
            return []

    except Exception as e:
        print(f"Error during scraping: {e}")
        return []

if __name__ == "__main__":
    # Test our scraper independently
    results = asyncio.run(scrape_jobs("Java Developer", "Germany"))
    print(f"Found {len(results)} jobs.")
