ai-engine/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
ollama
beautifulsoup4
requests
pydantic
pytest
